fix(tests): complete migration of implementation_manager to development-harness scripts
Migration commit 7893e088 updated pytest pythonpath to point at the
development-harness location but left the module files as symlinks
targeting the python3-development location instead of real files.
Convert the symlinks to real files in development-harness and delete
the now-redundant originals from python3-development.

Fixes #1063
